--- res.csv
@@ -1,29 +1,0 @@
-À,1225
-Á,1686
-Â,533
-Ã,1017
-Ä,782
-Å,106
-Æ,413
-Ç,645
-È,354
-É,4
-Ê,2321
-Ë,712
-Ì,1305
-Í,471
-Î,810
-Ï,1801
-Ð,588
-Ñ,1826
-Ò,1013
-Ó,272
-Ô,197
-Õ,288
-Ö,228
-×,692
-Ø,286
-Ù,159
-Ý,223
-Þ,138
-ß,211
